Initial commit of pwAlphaNumeric script
This commit adds just the letters required to spell out 'hello world'
into the pwAlphaNumeric.glf script which is meant to be used as a
'library'/'module' which other Glyph scripts can load/include and the
use to create individual letters/numbers.

Eventually, I will add all the rest of the letters and numbers plus
perhaps some common punctuation for fun.