Fix createComplexDomain proc
This commit fixes and simplifies the createComplexDomain procedure. I
had attempted to use the 'args' feature of TCL procs which supports a
variable number of arguments, (for the list of inner edges) however,
this is more easily done by simply passing a list of lists as the second
argument. This second argument defaults to an empty list to support
simple domains with no inner loops.